+++ src/runtime/components/BoosterVimeo/Base.vue
@@ -13,7 +13,11 @@
         @load="onLoad"
       />
       <default-button aria-label="Play Video" @click="onInit">
-        <component :is="pictureComponent" class="poster" v-bind="poster" />
+        <booster-picture
+          class="poster"
+          v-bind="pictureDataset"
+          :title="playerTitle"
+        />
         <slot v-if="loading" name="loading-spinner" />
         <slot v-if="!ready && !loading" name="play" />
       </default-button>
@@ -24,7 +28,7 @@
 </template>
 
 <script>
-import { useHead, markRaw, computed, ref } from '#imports';
+import { useHead, markRaw, computed, ref, useBoosterCritical } from '#imports';
 
 import DefaultButton from '../Button';
 import { load, ready } from './utils/loader';
@@ -69,30 +73,26 @@ export default {
       }
     },
 
-    posterSizes: {
-      type: Object,
+    posterSources: {
+      type: Array,
       default() {
-        return {
-          default: '100vw',
-          xxs: '100vw',
-          xs: '100vw',
-          sm: '100vw',
-          md: '100vw',
-          lg: '100vw',
-          xl: '100vw',
-          xxl: '100vw'
-        };
+        return [
+          {
+            src: undefined,
+            media: 'all',
+            sizes: {
+              default: '100vw'
+            }
+          }
+        ];
       }
-    },
-
-    posterDensities: {
-      type: [String, Number],
-      default: undefined
     }
   },
   emits: ['playing', 'ready'],
 
   async setup(props) {
+    useBoosterCritical();
+
     const script = ref([]);
 
     if (!(import.meta.server || process.env.prerender)) {
@@ -167,31 +167,19 @@ export default {
       };
     },
 
-    pictureComponent() {
-      return this.videoData ? BoosterPicture : 'picture';
-    },
-
-    poster() {
-      if (!this.videoData) {
-        return null;
-      }
+    pictureDataset() {
       return {
         formats: this.$booster.targetFormats,
-        title: this.playerTitle,
-        sources: [
-          {
-            format: 'jpg',
-            src:
-              this.videoData &&
-              this.videoData.thumbnail_url?.replace(
-                'https://i.vimeocdn.com',
-                'vimeo'
-              ),
-            sizes: this.posterSizes,
-            media: 'all',
-            densities: this.posterDensities
-          }
-        ]
+        title: this.title,
+        sources: this.posterSources.map(source => ({
+          ...source,
+          src:
+            source.src ||
+            this.videoData?.thumbnail_url?.replace(
+              'https://i.vimeocdn.com',
+              'vimeo'
+            )
+        }))
       };
     }
   },