+++ src/runtime/components/BoosterYoutube/Base.vue
@@ -20,7 +20,7 @@
 </template>
 
 <script>
-import { useHead, ref, computed, markRaw } from '#imports';
+import { useHead, ref, computed, markRaw, useBoosterCritical } from '#imports';
 
 import DefaultButton from '../Button';
 import { load } from './utils/loader';
@@ -69,31 +69,27 @@ export default {
       }
     },
 
-    posterSizes: {
-      type: Object,
+    posterSources: {
+      type: Array,
       default() {
-        return {
-          default: '100vw',
-          xxs: '100vw',
-          xs: '100vw',
-          sm: '100vw',
-          md: '100vw',
-          lg: '100vw',
-          xl: '100vw',
-          xxl: '100vw'
-        };
+        return [
+          {
+            src: undefined,
+            media: 'all',
+            sizes: {
+              default: '100vw'
+            }
+          }
+        ];
       }
-    },
-
-    posterDensities: {
-      type: [String, Number],
-      default: undefined
     }
   },
 
   emits: ['ready', 'playing'],
 
   setup() {
+    useBoosterCritical();
+
     const script = ref([]);
     useHead({
       script: computed(() => {
@@ -121,14 +117,10 @@ export default {
       return {
         formats: this.$booster.targetFormats,
         title: this.title,
-        sources: [
-          {
-            src: `/youtube/vi/${this.videoId}/maxresdefault.jpg`,
-            sizes: this.posterSizes,
-            media: 'all',
-            densities: this.posterDensities
-          }
-        ]
+        sources: this.posterSources.map(source => ({
+          ...source,
+          src: source.src || `/youtube/vi/${this.videoId}/maxresdefault.jpg`
+        }))
       };
     }
   },